moved DataPerDriver into its own module Removed _globals. This has been integrated into DriverData. Duplicate fields yet to be optimised
