+++ src/telemetry_data.py
@@ -25,16 +25,15 @@
 
 import copy
 import json
-from typing import Optional, Generator, Tuple, List, Dict, Any
+from typing import Optional, Generator, Tuple, List, Dict, Any, Union
 from readerwriterlock import rwlock
 from lib.f1_types import PacketSessionData, PacketLapData, LapData, CarTelemetryData, ParticipantData, \
     PacketEventData, PacketParticipantsData, PacketCarTelemetryData, PacketCarStatusData, FinalClassificationData, \
     PacketFinalClassificationData, PacketCarDamageData, PacketSessionHistoryData, ResultStatus, PacketTyreSetsData, \
     F1Utils, WeatherForecastSample, CarDamageData, CarStatusData, TrackID, ActualTyreCompound, \
     SafetyCarType, TelemetrySetting, PacketMotionData, CarMotionData, PacketCarSetupData, CarSetupData, ResultStatus, \
-    PacketTimeTrialData
-from src.data_per_driver import TyreSetInfo, TyreSetHistoryEntry, TyreInfo, \
-    PerLapSnapshotEntry, WarningPenaltyHistory, DriverInfo, LapInfo, PacketCopies, CarInfo
+    PacketTimeTrialData, SessionType23, SessionType24
+from src.data_per_driver import DataPerDriver
 from lib.race_analyzer import getFastestTimesJson, getTyreStintRecordsDict
 from lib.overtake_analyzer import OvertakeRecord
 from lib.collisions_analyzer import CollisionRecord, CollisionAnalyzer, CollisionAnalyzerMode
@@ -151,744 +150,6 @@ class GlobalData:
         self.m_safety_car_status = packet.m_safetyCarStatus
         return ret_status
 
-class DataPerDriver:
-    """
-    Class that models the data stored per race driver.
-
-    Attributes:
-        m_position (Optional[int]): The current position of the driver in the race.
-        m_name (Optional[str]): The name of the driver.
-        m_team (Optional[str]): The team to which the driver belongs.
-        m_driver_number (Optional[int]): The race number of the driver
-        m_delta_to_car_in_front (Optional[str]): The time difference between the driver and the car in front.
-        m_delta_to_leader (Optional[str]): The time difference to the race leader.
-        m_ers_perc (Optional[float]): The percentage of ERS (Energy Recovery System) remaining.
-        m_lap_info (LapInfo): Details regarding best and last lap
-        m_best_lap_ms (Optional[int]): The best lap time achieved by the driver.
-        m_last_lap_ms (Optional[int]): The time taken for the last lap completed by the driver in ms.
-        m_tyre_wear (Optional[TyreWearPerLap]): The level of wear on the driver's tires.
-        m_is_player (Optional[bool]): Indicates whether the driver is the player.
-        m_current_lap (Optional[int]): The current lap the driver is on.
-        m_penalties (Optional[str]): Penalties accumulated by the driver.
-        m_tyre_age (Optional[int]): The age of the driver's current set of tires.
-        m_tyre_vis_compound (Optional[VisualTyreCompound]): The visual type of tyre compound being used by the driver.
-        m_tyre_act_compound (Optional[ActualTyreCompound]): The actual type of tyre compound being used by the driver.
-        m_tyre_surface_temp (Optional[float]): The surface temperature of the driver's tires.
-        m_tyre_inner_temp (Optional[float]): The inner temperature of the driver's tires.
-        m_is_pitting (Optional[bool]): Indicates whether the driver is currently in the pit lane.
-        m_drs_activated (Optional[bool]): Indicates whether the DRS (Drag Reduction System) is activated for the driver.
-        m_drs_allowed (Optional[bool]): Indicates whether DRS is allowed for the driver.
-        m_drs_distance (Optional[int]): The distance to the car in front for DRS activation.
-        m_num_pitstops (Optional[int]): The number of pitstops made by the driver.
-        m_dnf_status_code (Optional[str]): Status code indicating if the driver did not finish the race.
-        m_tyre_life_remaining_laps (Optional[int]): The remaining laps the tires are expected to last.
-        m_telemetry_restrictions (Optional[TelemetrySetting]):
-            Telemetry settings indicating the level of data available for the driver.
-        m_tyre_set_history_manager (TyreSetHistoryManager)
-        m_tyre_wear_extrapolator (TyreWearExtrapolator): Predicts the tyre wear for upcoming laps
-        m_curr_lap_sc_status (SafetyCarStatus): The current lap's safety car status
-        m_fuel_load_kg (float): The current fuel load (in kg)
-        m_fuel_laps_remaining (float): Number of laps remaining with current fuel load
-        m_fl_wing_damage (int): Left front wing damage
-        m_fr_wing_damage (int): Right front wing damage
-        m_rear_wing_damage (int): Rear wing damage
-        m_result_status (Optional[ResultStatus]): The result status of the driver.
-        m_top_speed_kmph (Optional[float]): The top speed achieved by the driver.
-        m_collision_records (List[CollisionRecord]): List of CollisionRecord objects for the driver.
-        m_fuel_rate_recommender (Optional[FuelRateRecommender]): Fuel usage rate recommender for the driver.
-        self.m_warning_penalty_history: WarningPenaltyHistory
-
-        m_packet_lap_data (Optional[LapData]): Copy of LapData packet for the driver.
-        m_packet_participant_data (Optional[ParticipantData]): Copy of ParticipantData packet for the driver.
-        m_packet_car_telemetry (Optional[CarTelemetryData]): Copy of CarTelemetryData packet for the driver.
-        m_packet_car_status (Optional[CarStatusData]): Copy of CarStatusData packet for the driver.
-        m_packet_car_damage (Optional[CarDamageData]): Copy of CarDamageData packet for the driver.
-        m_packet_session_history (Optional[PacketSessionHistoryData]):
-            Copy of PacketSessionHistoryData packet for the driver.
-        m_packet_tyre_sets (Optional[PacketTyreSetsData]): Copy of PacketTyreSetsData packet for the driver.
-        m_packet_final_classification (Optional[FinalClassificationData]):
-            Copy of FinalClassificationData packet for the driver.
-    """
-
-    def __repr__(self) -> str:
-        """Get the string representation of this object
-
-        Returns:
-            str: The string representation
-        """
-
-        return f"DataPerDriver({self.m_driver_info.position}|{self.m_driver_info.name}|{self.m_driver_info.team})"
-
-    def __str__(self) -> str:
-        """Get the string representation of this object
-
-        Returns:
-            str: The string representation
-        """
-
-        return self.__repr__()
-
-    def __init__(self, total_laps):
-        """
-        Init the data per driver fields
-        """
-
-        self.m_driver_info: DriverInfo = DriverInfo()
-        self.m_lap_info: LapInfo = LapInfo()
-        self.m_tyre_info: TyreInfo = TyreInfo(total_laps)
-
-        self.m_car_info: CarInfo = CarInfo(total_laps)
-
-        self.m_collision_records: List[CollisionRecord] = []
-        self.m_warning_penalty_history: WarningPenaltyHistory = WarningPenaltyHistory()
-
-        # packet copies
-        self.m_packet_copies: PacketCopies = PacketCopies()
-
-        # Per lap snapshot
-        self.m_per_lap_snapshots: Dict[int, PerLapSnapshotEntry] = {}
-
-    @property
-    def is_valid(self) -> bool:
-        """Check if this DataPerDriver entry is valid. Reuse the same fields as __repr__
-
-        Returns:
-            bool: True if valid
-        """
-        return bool(self.m_driver_info.position or self.m_driver_info.name or self.m_driver_info.team)
-
-    def toJSON(self,
-               index: Optional[int] = None,
-               include_tyre_wear_prediction : Optional[bool] = False,
-               selected_pit_stop_lap : Optional[int] = None) -> Dict[str, Any]:
-        """Get a JSON representation of this DataPerDriver object
-
-        Args:
-            index (int): The index number. Defaults to None.
-            include_tyre_wear_prediction (Optional[bool]): Whether to include the tyre wear prediction
-            selected_pit_stop_lap (Optional[int]): The lap number of the selected pit stop
-
-        Returns:
-            Dict[str, Any]: The JSON dict
-        """
-        final_json = {}
-
-        # Add the primary data
-        if index is not None:
-            final_json["index"] = index
-        final_json["is-player"] = self.m_driver_info.is_player
-        final_json["driver-name"] = self.m_driver_info.name
-        final_json["track-position"] = self.m_driver_info.position
-        final_json["team"] = self.m_driver_info.team
-        final_json["telemetry-settings"] = str(self.m_driver_info.telemetry_restrictions)
-        final_json["current-lap"] = self.m_lap_info.m_current_lap
-        final_json["top-speed-kmph"] = self.m_lap_info.m_top_speed_kmph_this_lap
-
-        # Insert packet copies if available
-        final_json["car-damage"] = self.m_packet_copies.m_packet_car_damage.toJSON() if self.m_packet_copies.m_packet_car_damage else None
-        final_json["car-status"] = self.m_packet_copies.m_packet_car_status.toJSON() if self.m_packet_copies.m_packet_car_status else None
-        final_json["participant-data"] = self.m_packet_copies.m_packet_particpant_data.toJSON() if self.m_packet_copies.m_packet_particpant_data else None
-        final_json["tyre-sets"] = self.m_packet_copies.m_packet_tyre_sets.toJSON() if self.m_packet_copies.m_packet_tyre_sets else None
-        final_json["session-history"] = self.m_packet_copies.m_packet_session_history.toJSON() if self.m_packet_copies.m_packet_session_history else None
-        final_json["final-classification"] = self.m_packet_copies.m_packet_final_classification.toJSON() if self.m_packet_copies.m_packet_final_classification else None
-        final_json["lap-data"] = self.m_packet_copies.m_packet_lap_data.toJSON() if self.m_packet_copies.m_packet_lap_data else None
-        final_json["car-setup"] = self.m_packet_copies.m_packet_car_setup.toJSON() if self.m_packet_copies.m_packet_car_setup else None
-        final_json["warning-penalty-history"] = [entry.toJSON() for entry in self.m_warning_penalty_history.getEntries()]
-
-        # Insert the tyre set history
-        final_json["tyre-set-history"]= self._getTyreSetHistoryJSON()
-
-        # Insert the per lap snapshot
-        final_json["per-lap-info"] = []
-        for lap_number, snapshot_entry in self._getNextLapSnapshot():
-            final_json["per-lap-info"].append(snapshot_entry.toJSON(lap_number))
-
-        if include_tyre_wear_prediction:
-            final_json["tyre-wear-predictions"] = self.getFullTyreWearPredictions(selected_pit_stop_lap)
-
-        # Insert the lap time history against tyre used
-        final_json["lap-time-history"] = self._getLapTimeHistoryJSON() # can be None, handled in frontend
-
-        # Collisions data
-        final_json["collisions"] = self.getCollisionStatsJSON()
-
-        # Return this fully prepped JSON
-        return final_json
-
-    def getFullTyreWearPredictions(self, selected_pit_stop_lap: Optional[int] = None) -> Dict[str, Any]:
-        """Get a JSON list with the tyre wear predictions for all remaining laps
-
-        Args:
-            next_pit_window (Optional[int], optional): The next pit window lap number.
-
-        Returns:
-            List[Dict[str, Any]]: List of JSON objects, each containing tyre wear predictions for a specific lap
-        """
-
-        if self.m_tyre_info.m_tyre_wear_extrapolator.isDataSufficient():
-            return {
-                "status" : True,
-                "desc" : "Data is sufficient for extrapolation",
-                "predictions": [item.toJSON() for item in self.m_tyre_info.m_tyre_wear_extrapolator.predicted_tyre_wear],
-                "selected-pit-stop-lap": selected_pit_stop_lap
-            }
-        else:
-            return {
-                "status" : False,
-                "desc" : "Insufficient data for extrapolation",
-                "predictions": [],
-                "selected-pit-stop-lap": None
-            }
-
-    def getTyrePredictionsJSONList(self, next_pit_window: Optional[int] = None) -> List[Dict[str, Any]]:
-        """Get a JSON list with the tyre wear predictions for next stop/mid point and end of race
-
-        Args:
-            next_pit_window (Optional[int], optional): The next pit window lap number.
-                If None, then returns the predictions for final lap and half way through between current and final lap
-
-        Returns:
-            List[Dict[str, Any]]: List of JSON objects, each containing tyre wear predictions for a specific lap
-        """
-
-        if not self.m_tyre_info.m_tyre_wear_extrapolator.isDataSufficient():
-            # Data unavailable, return empty list
-            return []
-
-        predictions_list = []
-
-        # Input sanitization
-        if next_pit_window is None or (next_pit_window == 0) or (next_pit_window < self.m_lap_info.m_current_lap):
-            # Lets return the lap midway between current lap and final lap
-            next_pit_window = (self.m_lap_info.m_current_lap + self.m_tyre_info.m_tyre_wear_extrapolator.total_laps) // 2
-
-        # NOTE: Flashbacks or delayed telemetry starts can cause the prediction to not be available.
-        # This happens in the last lap
-        if next_pit_window == self.m_tyre_info.m_tyre_wear_extrapolator.total_laps:
-            # We are already in the final lap, so return the final prediction
-            predicted_tyre_wear = self.m_tyre_info.m_tyre_wear_extrapolator.getTyreWearPrediction()
-            if predicted_tyre_wear:
-                predictions_list.append(predicted_tyre_wear.toJSON())
-        else:
-
-            # Add prediction for next window if available
-            pit_lap_prediction = self.m_tyre_info.m_tyre_wear_extrapolator.getTyreWearPrediction(next_pit_window)
-            if pit_lap_prediction and next_pit_window:
-                predictions_list.append(pit_lap_prediction.toJSON())
-
-            # Add final lap prediction if available
-            final_lap_prediction = self.m_tyre_info.m_tyre_wear_extrapolator.getTyreWearPrediction()
-            if final_lap_prediction:
-                predictions_list.append(final_lap_prediction.toJSON())
-        return predictions_list
-
-    def getCurrentTyreWearJSON(self) -> Dict[str, Any]:
-        """Get the current tyre wear in JSON format
-
-        Returns:
-            JSON object: JSON object containing the current tyre wear
-        """
-
-        return self.m_tyre_info.tyre_wear.toJSON() if self.m_tyre_info.tyre_wear else None
-
-    def _getTyreSetHistoryJSON(self, include_wear_history: Optional[bool] = True) -> List[Dict[str, Any]]:
-        """Get the list of tyre sets used in JSON format
-
-        Args:
-            include_wear_history (Optional[bool]): Whether to include the tyre wear history. Defaults to True.
-
-        Returns:
-            JSON list: JSON list containing multiple JSON objects, each representing one set of tyres used, in order.
-        """
-
-        return self.m_tyre_info.m_tyre_set_history_manager.toJSON(include_wear_history,
-                                                    self.m_packet_copies.m_packet_tyre_sets,
-                                                    self.m_packet_copies.m_packet_session_history.m_numLaps)
-
-    def _getTyreWearHistoryJSON(self, start_lap : int, end_lap : int):
-        """
-        Generate JSON data for tyre wear history within specified lap range.
-
-        Args:
-            start_lap (int): The starting lap number.
-            end_lap (int): The ending lap number.
-
-        Returns:
-            list: A list of dictionaries containing lap number and tyre wear data.
-        """
-
-        range_of_laps = range(start_lap, end_lap + 1)
-        tyre_wear_history = []
-        if start_lap == 1:
-            tyre_wear_history.append({
-                'lap-number': 0,
-                'front-right-wear': self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_FRONT_RIGHT],
-                'front-left-wear': self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_FRONT_LEFT],
-                'rear-right-wear': self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_REAR_RIGHT],
-                'rear-left-wear': self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_REAR_LEFT],
-            })
-        for lap_number in range_of_laps:
-            if lap_number in self.m_per_lap_snapshots:
-                car_damage_data = self.m_per_lap_snapshots[lap_number].m_car_damage_packet
-                if car_damage_data:
-                    tyre_wear_history.append({
-                        'lap-number': lap_number,
-                        'front-right-wear': car_damage_data.m_tyresWear[F1Utils.INDEX_FRONT_RIGHT],
-                        'front-left-wear': car_damage_data.m_tyresWear[F1Utils.INDEX_FRONT_LEFT],
-                        'rear-right-wear': car_damage_data.m_tyresWear[F1Utils.INDEX_REAR_RIGHT],
-                        'rear-left-wear': car_damage_data.m_tyresWear[F1Utils.INDEX_REAR_LEFT],
-                    })
-                else:
-                    png_logger.debug('car damage data not available for lap %d driver %s', lap_number,
-                        self.m_driver_info.name)
-            else:
-                png_logger.debug('per lap snapshot not available for lap %d driver %s', lap_number,
-                    self.m_driver_info.name)
-        return tyre_wear_history
-
-    def _getLapTimeHistoryJSON(self) -> Dict[str, Any]:
-        """Get the lap time history in JSON format
-
-        Returns:
-            JSON object: JSON object containing the lap time history and tyre info for each lap
-        """
-
-        if not self.m_packet_copies.m_packet_session_history:
-            return None
-        per_lap_tyre_info = self._getPerLapTyreInfoJSON()
-        lap_history_data = []
-        for index, entry in enumerate(self.m_packet_copies.m_packet_session_history.m_lapHistoryData):
-            # Get tyre set history at start of lap (i.e.) end of previous lap
-            lap_number = index # Use index since it is lap_number - 1
-            # Find the tyre set at the specified lap
-            tyre_set_info = next((obj for obj in per_lap_tyre_info if obj.get("lap-number") == lap_number), None)
-            lap_history_data.append({
-                "lap-time-in-ms": entry.m_lapTimeInMS,
-                "lap-time-str": F1Utils.millisecondsToMinutesSecondsMilliseconds(entry.m_lapTimeInMS),
-                "sector-1-time-in-ms": entry.m_sector1TimeInMS,
-                "sector-1-time-minutes": entry.m_sector1TimeMinutes,
-                "sector-1-time-str" : F1Utils.getLapTimeStrSplit(entry.m_sector1TimeMinutes,
-                    entry.m_sector1TimeInMS),
-                "sector-2-time-in-ms": entry.m_sector2TimeInMS,
-                "sector-2-time-minutes": entry.m_sector2TimeMinutes,
-                "sector-2-time-str": F1Utils.getLapTimeStrSplit(entry.m_sector2TimeMinutes,
-                    entry.m_sector2TimeInMS),
-                "sector-3-time-in-ms": entry.m_sector3TimeInMS,
-                "sector-3-time-minutes": entry.m_sector3TimeMinutes,
-                "sector-3-time-str": F1Utils.getLapTimeStrSplit(entry.m_sector3TimeMinutes,
-                    entry.m_sector3TimeInMS),
-                "lap-valid-bit-flags": entry.m_lapValidBitFlags,
-                "tyre-set-info" : tyre_set_info,
-            })
-        return {
-            "best-lap-time-lap-num": self.m_packet_copies.m_packet_session_history.m_bestLapTimeLapNum,
-            "best-sector-1-lap-num": self.m_packet_copies.m_packet_session_history.m_bestSector1LapNum,
-            "best-sector-2-lap-num": self.m_packet_copies.m_packet_session_history.m_bestSector2LapNum,
-            "best-sector-3-lap-num": self.m_packet_copies.m_packet_session_history.m_bestSector3LapNum,
-            "lap-history-data": lap_history_data,
-        }
-
-    def _getPerLapTyreInfoJSON(self) -> List[Dict[str, Any]]:
-        """Get the per lap tyre info in JSON format
-
-        Returns:
-            JSON list: JSON list containing multiple JSON objects, each representing one lap time, in order.
-        """
-
-        ret = []
-        for tyre_set_meta_data in self.m_tyre_info.m_tyre_set_history_manager.getEntries():
-            for tyre_wear in tyre_set_meta_data.m_tyre_wear_history:
-                tyre_set_data = self.m_per_lap_snapshots[tyre_wear.lap_number] \
-                    .m_tyre_sets_packet.m_tyreSetData[tyre_set_meta_data.m_fitted_index] \
-                            if tyre_wear.lap_number in self.m_per_lap_snapshots else None
-                ret.append({
-                    'tyre-wear' : tyre_wear.toJSON(),
-                    'lap-number' : tyre_wear.lap_number,
-                    'tyre-set' : tyre_set_data.toJSON() if tyre_set_data else None,
-                })
-        return ret
-
-    def onLapChange(self,
-        old_lap_number: int) -> None:
-        """
-        Perform snapshot for the given lap change.
-
-        Args:
-            old_lap_number (int): The old lap number.
-        """
-
-        # Check if the old lap number is already present in the snapshots (lap already processed)
-        if old_lap_number in self.m_per_lap_snapshots:
-            return
-
-        # Store the snapshot data for the old lap
-        self.m_per_lap_snapshots[old_lap_number] = PerLapSnapshotEntry(
-            car_damage=self.m_packet_copies.m_packet_car_damage,
-            car_status=self.m_packet_copies.m_packet_car_status,
-            sc_status=self.m_driver_info.m_curr_lap_sc_status,
-            tyre_sets=self.m_packet_copies.m_packet_tyre_sets,
-            track_position=self.m_driver_info.position,
-            top_speed_kmph=self.m_lap_info.m_top_speed_kmph_this_lap,
-        )
-
-        # Now clear the top speed
-        self.m_lap_info.m_top_speed_kmph_this_lap = None
-
-        # Add the tyre wear data into the tyre stint history
-        if old_lap_number and self.m_tyre_info.m_tyre_set_history_manager.length:
-            self.m_tyre_info.m_tyre_set_history_manager.addTyreWear(TyreWearPerLap(
-                fl_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_FRONT_LEFT],
-                fr_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_FRONT_RIGHT],
-                rl_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_REAR_LEFT],
-                rr_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_REAR_RIGHT],
-                lap_number=old_lap_number,
-                is_racing_lap=True,
-                desc=f"end of lap {old_lap_number} snapshot"
-            ))
-
-        # Add the tyre wear data into the extrapolator
-        tyre_set_id = self._getCurrentTyreSetKey()
-        if tyre_set_id:
-            self.m_tyre_info.m_tyre_wear_extrapolator.updateDataLap(TyreWearPerLap(
-                fl_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_FRONT_LEFT],
-                fr_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_FRONT_RIGHT],
-                rl_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_REAR_LEFT],
-                rr_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_REAR_RIGHT],
-                lap_number=old_lap_number,
-                is_racing_lap=True,
-                desc=tyre_set_id
-            ))
-
-        # Fuel stuff
-        if self.m_packet_copies.m_packet_car_status:
-            self.m_car_info.m_fuel_rate_recommender.add(self.m_packet_copies.m_packet_car_status.m_fuelInTank, old_lap_number)
-
-    def isZerothLapSnapshotDataAvailable(self) -> bool:
-        """
-        Checks if zeroth lap snapshot data is available.
-
-        Returns:
-            bool - True if zeroth lap snapshot data is available
-        """
-
-        return bool(
-            self.m_packet_copies.m_packet_car_damage and
-            self.m_packet_copies.m_packet_car_status and
-            self.m_packet_copies.m_packet_tyre_sets and
-            self.m_driver_info.position and
-            (self.m_lap_info.m_top_speed_kmph_this_lap is not None) # TODO: is this required for zeroth lap?
-        )
-
-    def updateTyreSetData(self, fitted_index: int) -> None:
-        """Update the current tyre set in the history list, if required.
-               NOTE: The tyre history is ignored if the player has disabled telemetry
-
-        Args:
-            fitted_index (int): The fitted tyre set index
-        """
-
-        # fuck those anti telemetry cunts
-        if self.m_driver_info.telemetry_restrictions != TelemetrySetting.PUBLIC:
-            return
-
-        # This can happen if tyre sets packets arrives before lap data packet
-        if self.m_lap_info.m_current_lap is not None:
-            fitted_tyre_set_key = self._getCurrentTyreSetKey()
-            if not self.m_tyre_info.m_tyre_set_history_manager.length:
-                if 0 in self.m_per_lap_snapshots:
-                    # Start of race, enter the tyre wear data along with starting value
-                    initial_tyre_wear = TyreWearPerLap(
-                        fl_tyre_wear=self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_FRONT_LEFT],
-                        fr_tyre_wear=self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_FRONT_RIGHT],
-                        rl_tyre_wear=self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_REAR_LEFT],
-                        rr_tyre_wear=self.m_per_lap_snapshots[0].m_car_damage_packet.m_tyresWear[F1Utils.INDEX_REAR_RIGHT],
-                        lap_number=0,
-                        is_racing_lap=True,
-                        desc="end of zeroth lap data point"
-                    )
-                    self.m_tyre_info.m_tyre_set_history_manager.add(TyreSetHistoryEntry(
-                                                start_lap=self.m_lap_info.m_current_lap,
-                                                index=fitted_index,
-                                                tyre_set_key=fitted_tyre_set_key,
-                                                initial_tyre_wear=initial_tyre_wear,
-                    ))
-            elif fitted_index != self.m_tyre_info.m_tyre_set_history_manager.getLastEntry().m_fitted_index:
-                lap_number = self.m_lap_info.m_current_lap - 1
-                # create a new tyre set entry with initial data.
-                initial_tyre_wear = TyreWearPerLap(
-                    fl_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_FRONT_LEFT],
-                    fr_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_FRONT_RIGHT],
-                    rl_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_REAR_LEFT],
-                    rr_tyre_wear=self.m_packet_copies.m_packet_car_damage.m_tyresWear[F1Utils.INDEX_REAR_RIGHT],
-                    lap_number=lap_number,
-                    is_racing_lap=True,
-                    desc=f"tyre set change detected. key={str(fitted_tyre_set_key)}"
-                )
-                self.m_tyre_info.m_tyre_set_history_manager.add(TyreSetHistoryEntry(
-                                            start_lap=lap_number,
-                                            index=fitted_index,
-                                            tyre_set_key=fitted_tyre_set_key,
-                                            initial_tyre_wear=initial_tyre_wear,
-                ))
-
-                # Tyre set change detected. clear the extrapolation data
-                self.m_tyre_info.m_tyre_wear_extrapolator.clear()
-                self.m_tyre_info.m_tyre_wear_extrapolator.updateDataLap(initial_tyre_wear)
-
-    def _getCurrentTyreSetKey(self) -> Optional[str]:
-        """Get the unique ID key for the currently equipped tyre set
-
-        Returns:
-            Optional[str]: The tyre set key
-        """
-
-        return self.m_packet_copies.m_packet_tyre_sets.getFittedTyreSetKey() \
-            if self.m_packet_copies.m_packet_tyre_sets else None
-
-    def _getCurrentTyreSetID(self) -> Optional[int]:
-        """Get the ID/index for the currently equipped tyre set
-
-        Returns:
-            Optional[int]: The tyre set ID
-        """
-
-        return self.m_packet_copies.m_packet_tyre_sets.m_fittedIdx if self.m_packet_copies.m_packet_tyre_sets else None
-
-    def _getNextLapSnapshot(self) -> Generator[Tuple[int, PerLapSnapshotEntry], None, None]:
-        """
-        Returns a generator for each lap's snapshot in order.
-
-        Yields:
-            Tuple[int, PerLapHistoryEntry]: Tuple containing lap number and snapshot data for each lap.
-        """
-        for lap_number in sorted(self.m_per_lap_snapshots.keys()):
-            yield lap_number, self.m_per_lap_snapshots[lap_number]
-
-    def getTyreSetInfoAtLap(self, lap_num: Optional[int] = None) -> Optional[TyreSetInfo]:
-        """Get the tyre set info at the specified lap number
-
-        Args:
-            lap_num (Optional[int]): The lap number. If None, uses current lap number
-
-        Returns:
-            Optional[TyreSetInfo]: The tyre set info. None if data not found or invalid lap num
-        """
-
-        if self.m_lap_info.m_current_lap is None:
-            return None
-
-        if lap_num is None:
-            lap_num = self.m_lap_info.m_current_lap
-
-        if lap_num == self.m_lap_info.m_current_lap and self.m_packet_copies.m_packet_car_status:
-            return TyreSetInfo(
-                actual_tyre_compound=self.m_packet_copies.m_packet_car_status.m_actualTyreCompound,
-                visual_tyre_compound=self.m_packet_copies.m_packet_car_status.m_visualTyreCompound,
-                tyre_set_id=self._getCurrentTyreSetID(),
-                tyre_age_laps=self.m_packet_copies.m_packet_car_status.m_tyresAgeLaps
-            )
-        if (lap_num < self.m_lap_info.m_current_lap) and (lap_num in self.m_per_lap_snapshots):
-            snapshot_at_lap       = self.m_per_lap_snapshots[lap_num]
-            snapshot_car_status   = snapshot_at_lap.m_car_status_packet
-            snapshot_tyre_sets    = snapshot_at_lap.m_tyre_sets_packet
-            if not snapshot_car_status or not snapshot_tyre_sets:
-                return None
-            return TyreSetInfo(
-                actual_tyre_compound=snapshot_car_status.m_actualTyreCompound,
-                visual_tyre_compound=snapshot_car_status.m_visualTyreCompound,
-                tyre_set_id=snapshot_tyre_sets.m_fittedIdx,
-                tyre_age_laps=snapshot_car_status.m_tyresAgeLaps
-            )
-
-        return None
-
-    def getCollisionStatsJSON(self) -> Dict[str, Any]:
-        """Get the collision stats JSON.
-
-        Returns:
-            Dict[str, Any]: Collision stats JSON
-        """
-
-        collision_analyzer = CollisionAnalyzer(
-            input_mode=CollisionAnalyzerMode.INPUT_MODE_LIST_COLLISION_RECORDS,
-            input_data=self.m_collision_records)
-        return collision_analyzer.toJSON()
-
-    def getFuelStatsJSON(self) -> Dict[str, Any]:
-        """Get the fuel stats JSON.
-
-        Returns:
-            Dict[str, Any]: Fuel stats JSON
-        """
-
-        if self.m_packet_copies.m_packet_car_status:
-            return {
-                "fuel-capacity" : self.m_packet_copies.m_packet_car_status.m_fuelCapacity,
-                "fuel-mix" : str(self.m_packet_copies.m_packet_car_status.m_fuelMix),
-                "fuel-remaining-laps" : self.m_packet_copies.m_packet_car_status.m_fuelRemainingLaps,
-                "fuel-in-tank" : self.m_packet_copies.m_packet_car_status.m_fuelInTank,
-                "curr-fuel-rate" :self.m_car_info.m_fuel_rate_recommender.curr_fuel_rate,
-                "last-lap-fuel-used" : self.m_car_info.m_fuel_rate_recommender.fuel_used_last_lap,
-                "target-fuel-rate-average" : self.m_car_info.m_fuel_rate_recommender.target_fuel_rate,
-                "target-fuel-rate-next-lap" : self.m_car_info.m_fuel_rate_recommender.target_next_lap_fuel_usage,
-                "surplus-laps" : self.m_car_info.m_fuel_rate_recommender.surplus_laps,
-            }
-
-        return {
-            "fuel-capacity" : 0.0,
-            "fuel-mix" : 0,
-            "fuel-remaining-laps" : 0.0,
-            "fuel-in-tank" : 0.0,
-            "curr-fuel-rate" : 0.0,
-            "last-lap-fuel-used" : 0.0,
-            "target-fuel-rate-average" : 0.0,
-            "target-fuel-rate-next-lap" : 0.0,
-            "surplus-laps" : 0.0,
-        }
-
-    def getPositionHistoryJSON(self) -> Dict[str, Any]:
-        """Get the position history JSON.
-
-        Returns:
-            Dict[str, Any]: Position history JSON
-        """
-
-        return {
-            "name": self.m_driver_info.name,
-            "team": self.m_driver_info.team,
-            "driver-number": self.m_driver_info.driver_number,
-            "driver-position-history": [
-                {
-                    "lap-number": lap_number,
-                    "position": snapshot_record.m_track_position
-                }
-                for lap_number, snapshot_record in self._getNextLapSnapshot()
-            ]
-        }
-
-    def getTyreStintHistoryJSON(self) -> Dict[str, Any]:
-        """Get the tyre stint history JSON.
-
-        Returns:
-            Dict[str, Any]: Tyre stint history JSON
-        """
-
-        return {
-            "name": self.m_driver_info.name,
-            "team": self.m_driver_info.team,
-            "driver-number": self.m_driver_info.driver_number,
-            "delta-to-leader" : self.m_lap_info.m_delta_to_leader,
-            "tyre-stint-history": self._getTyreSetHistoryJSON(include_wear_history=False),
-        }
-
-    def updateLapDataPacketCopy(self, lap_data: LapData, full_lap_distance: int) -> None:
-        """Add to the warning/penalty history if required and update the lap data packet copy
-
-        Args:
-            lap_data (LapData): The incoming lap data packet
-            full_lap_distance (int): The distance of the entire lap in metres
-        """
-
-        # Update the history and packet copy
-        self.m_warning_penalty_history.update(lap_data, full_lap_distance, self.m_packet_copies.m_packet_lap_data)
-        self.m_packet_copies.m_packet_lap_data = lap_data
-
-    def getSectorStatus(self,
-                        sector_1_best_ms: Optional[int],
-                        sector_2_best_ms: Optional[int],
-                        sector_3_best_ms: Optional[int],
-                        for_best_lap: bool,
-                        session_type_str: str) -> List[Optional[int]]:
-        # sourcery skip: merge-duplicate-blocks, remove-redundant-if
-        """
-        Determine sector status for either best or last lap.
-
-        Args:
-            sector_1_best_ms: Best sector 1 time in milliseconds
-            sector_2_best_ms: Best sector 2 time in milliseconds
-            sector_3_best_ms: Best sector 3 time in milliseconds
-            for_best_lap: Whether to analyze best lap or last lap
-            session_type_str: The session type string
-
-        Returns:
-            List of sector statuses (purple, green, yellow, invalid, or NA)
-        """
-        default_val = [
-            F1Utils.SECTOR_STATUS_NA,
-            F1Utils.SECTOR_STATUS_NA,
-            F1Utils.SECTOR_STATUS_NA
-        ]
-
-        # Validate input/reference data
-        if (not self.m_lap_info.m_pb_s1_ms or
-            not self.m_lap_info.m_pb_s2_ms or
-            not self.m_lap_info.m_pb_s3_ms or
-            not sector_1_best_ms or
-            not sector_2_best_ms or
-            not sector_3_best_ms):
-            return default_val
-
-        # Validate driver's data. For best/last lap, all relevant fields must be present
-        if for_best_lap and not self.m_lap_info.m_best_lap_obj:
-            return default_val
-        elif not self.m_lap_info.m_last_lap_obj:
-            return default_val
-
-        # Select lap object
-        lap_obj = self.m_lap_info.m_best_lap_obj if for_best_lap else self.m_lap_info.m_last_lap_obj
-        if not lap_obj:
-            return default_val
-
-        return [
-            self._get_sector_status(
-                sector_time=lap_obj.s1TimeMS,
-                sector_best_ms=sector_1_best_ms,
-                is_personal_best_sector_lap=(lap_obj.s1TimeMS == self.m_lap_info.m_pb_s1_ms),
-                sector_valid_flag=lap_obj.isSector1Valid()),
-            self._get_sector_status(
-                sector_time=lap_obj.s2TimeMS,
-                sector_best_ms=sector_2_best_ms,
-                is_personal_best_sector_lap=(lap_obj.s2TimeMS == self.m_lap_info.m_pb_s2_ms),
-                sector_valid_flag=lap_obj.isSector2Valid()),
-            self._get_sector_status(
-                sector_time=lap_obj.s3TimeMS,
-                sector_best_ms=sector_3_best_ms,
-                is_personal_best_sector_lap=(lap_obj.s3TimeMS == self.m_lap_info.m_pb_s3_ms),
-                sector_valid_flag=lap_obj.isSector3Valid()),
-        ] if lap_obj else default_val
-
-    def _get_sector_status(
-        self,
-        sector_time: int,
-        sector_best_ms: int,
-        is_personal_best_sector_lap: bool,
-        sector_valid_flag: bool
-    ) -> int:
-        """
-        Determine the status of a single sector.
-
-        Args:
-            sector_time: Time of the current sector
-            sector_best_ms: Best time for the sector
-            is_personal_best_sector_lap: Whether this is the best lap for this sector
-            sector_valid_flag: Whether the sector is valid
-
-        Returns:
-            Sector status (purple, green, yellow, invalid)
-        """
-        if sector_time == sector_best_ms:
-            # Session best
-            return F1Utils.SECTOR_STATUS_PURPLE
-        elif is_personal_best_sector_lap:
-            # Personal best
-            return F1Utils.SECTOR_STATUS_GREEN
-        elif not sector_valid_flag:
-            # Invalidated
-            return F1Utils.SECTOR_STATUS_INVALID
-        else:
-            # Meh sector
-            return F1Utils.SECTOR_STATUS_YELLOW
-
 class DriverData:
 
     """
@@ -923,11 +184,12 @@ class DriverData:
         self.m_num_active_cars: Optional[int] = None
         self.m_num_dnf_cars: Optional[int] = None
         self.m_race_completed: Optional[bool] = None
-        self.m_final_json: Dict[str, Any] = None
+        self.m_final_json: Dict[str, Any] = None # TODO: revisit
         self.m_is_player_dnf : Optional[bool] = None
         self.m_total_laps : Optional[int] = None
         self.m_ideal_pit_stop_window : Optional[int] = None
         self.m_track_id : Optional[TrackID] = None
+        self.m_session_type: Union[SessionType23 | SessionType24] = None
         self.m_game_year : Optional[int] = None
         self.m_collision_records : List[CollisionRecord] = []
         self.m_track_length : Optional[int] = None
@@ -935,6 +197,7 @@ class DriverData:
         self.m_fastest_s2_ms: Optional[int] = None
         self.m_fastest_s3_ms: Optional[int] = None
         self.m_time_trial_packet : Optional[PacketTimeTrialData] = None
+        self.m_globals: GlobalData = GlobalData()
 
     def clear(self) -> None:
         """Clear this object. Clears the m_driver_data list and sets everything else to None
@@ -950,12 +213,14 @@ class DriverData:
         self.m_total_laps = None
         self.m_ideal_pit_stop_window = None
         self.m_track_id = None
+        self.m_session_type = None
         self.m_game_year = None
         self.m_collision_records.clear()
         self.m_track_length = None
         self.m_fastest_s1_ms = None
         self.m_fastest_s2_ms = None
         self.m_fastest_s3_ms = None
+        self.m_globals.clear()
 
     def setRaceOngoing(self) -> None:
         """
@@ -1098,6 +363,7 @@ class DriverData:
 
         self.m_ideal_pit_stop_window = packet.m_pitStopWindowIdealLap
         self.m_track_id = packet.m_trackId
+        self.m_session_type = packet.m_sessionType
         self.m_game_year = packet.m_header.m_gameYear
         self.m_track_length = packet.m_trackLength
 
@@ -1269,7 +535,7 @@ class DriverData:
         """
 
         final_json = packet.toJSON()
-        is_position_history_supported = isPositionHistorySupported()
+        is_position_history_supported = self.isPositionHistorySupported()
         if is_position_history_supported:
             final_json["position-history"] = []
             final_json["tyre-stint-history"] = []
@@ -1289,6 +555,11 @@ class DriverData:
         final_json['classification-data'] = sorted(final_json['classification-data'], key=lambda x: x['track-position'])
         final_json['game-year'] = self.m_game_year
         self.m_final_json = final_json
+
+        final_json["session-info"] = self.m_globals.m_packet_session.toJSON() \
+            if self.m_globals.m_packet_session else None
+        self.m_globals.m_packet_final_classification = packet
+
         return final_json
 
     def processCarDamageUpdate(self, packet: PacketCarDamageData) -> None:
@@ -1441,6 +712,7 @@ class DriverData:
             return None
         final_json = driver_info_obj.toJSON(index, include_wear_prediction, selected_pit_stop_lap)
         final_json["circuit"] = str(self.m_track_id)
+        final_json["session-type"] = str(self.m_session_type)
         final_json["is-finish-line-after-pit-garage"] = F1Utils.isFinishLineAfterPitGarage(self.m_track_id) \
             if self.m_track_id is not None else None
 
@@ -1556,11 +828,17 @@ class DriverData:
             None,
         )
 
+    def isPositionHistorySupported(self) -> bool:
+        """Returns whether the position history is supported for the given event type
+            Position history is only supported in race events
+
+        Returns:
+            bool: True if the position history is supported, False otherwise
+        """
+        return bool(self.m_globals.m_event_type and "Race" in self.m_globals.m_event_type)
 # -------------------------------------- GLOBALS -----------------------------------------------------------------------
 
-_globals = GlobalData()
 _driver_data = DriverData()
-_globals_lock: rwlock.RWLockFair = rwlock.RWLockFair()
 _driver_data_lock: rwlock.RWLockFair = rwlock.RWLockFair()
 png_logger = getLogger()
 _custom_markers_history = CustomMarkersHistory()
@@ -1585,8 +863,7 @@ def processSessionUpdate(packet: PacketSessionData) -> bool:
 
     with _driver_data_lock.gen_wlock():
         _driver_data.processSessionUpdate(packet)
-    with _globals_lock.gen_wlock():
-        should_clear = _globals.processSessionUpdate(packet)
+        should_clear = _driver_data.m_globals.processSessionUpdate(packet)
     if should_clear:
         clearDataStructures("session update")
     return should_clear
@@ -1679,9 +956,6 @@ def processFinalClassificationUpdate(packet: PacketFinalClassificationData) -> D
     with _driver_data_lock.gen_wlock():
         final_json = _driver_data.processFinalClassificationUpdate(packet)
         _driver_data.setRaceCompleted()
-    with _globals_lock.gen_wlock():
-        final_json["session-info"] = _globals.m_packet_session.toJSON() if _globals.m_packet_session else None
-        _globals.m_packet_final_classification = packet
     final_json['custom-markers'] = _custom_markers_history.getJSONList()
     return final_json
 
@@ -1781,8 +1055,8 @@ def getGlobals() -> GlobalData:
     Returns:
         GlobalData: A copy of the GlobalData object
     """
-    with _globals_lock.gen_rlock():
-        return copy.deepcopy(_globals)
+    with _driver_data_lock.gen_rlock():
+        return copy.deepcopy(_driver_data.m_globals)
 
 def getDriverInfoJsonByIndex(index: int) -> Optional[Dict[str, Any]]:
     """Get the driver info JSON for the given index
@@ -1833,9 +1107,13 @@ def getEventInfoStr() -> Optional[str]:
     Returns:
         Optional[str]: The event type string (ends with an underscore), or None if no data is available
     """
-    with _globals_lock.gen_rlock():
-        if _globals.m_event_type and _globals.m_circuit:
-            return f"{_globals.m_event_type}_{_globals.m_circuit}".replace(' ', '_') + '_'
+    # with _globals_lock.gen_rlock():
+    #     if _globals.m_event_type and _globals.m_circuit:
+    #         return f"{_globals.m_event_type}_{_globals.m_circuit}".replace(' ', '_') + '_'
+    #     return None
+    with _driver_data_lock.gen_rlock():
+        if _driver_data.m_track_id and _driver_data.m_session_type:
+            return f"{str(_driver_data.m_session_type)}_{str(_driver_data.m_track_id)}".replace(' ', '_') + '_'
         return None
 
 def getOvertakeObj(overtaking_car_index: int, being_overtaken_index: int) -> Optional[OvertakeRecord]:
@@ -1882,30 +1160,29 @@ def getCustomMarkerEntryObj() -> Optional[CustomMarkerEntry]:
     """
 
     with _driver_data_lock.gen_rlock():
-        player_data = _driver_data.m_driver_data.get(_driver_data.m_player_index, None)
+        player_data = _driver_data.m_driver_data.get(_driver_data.m_player_index)
         if player_data:
             # CSV string - <track>,<event-type>,<lap-num>,<sector-num>
-            lap_num = player_data.m_current_lap
-            sector = player_data.m_packet_lap_data.m_sector
+            lap_num = player_data.m_lap_info.m_current_lap
+            sector = player_data.m_packet_copies.m_packet_lap_data.m_sector
             curr_lap_time = F1Utils.millisecondsToMinutesSecondsMilliseconds(
-                player_data.m_packet_lap_data.m_currentLapTimeInMS)
-            curr_lap_dist = player_data.m_packet_lap_data.m_lapDistance
+                player_data.m_packet_copies.m_packet_lap_data.m_currentLapTimeInMS)
+            curr_lap_dist = player_data.m_packet_copies.m_packet_lap_data.m_lapDistance
+
+            track = str(_driver_data.m_track_id)
+            event_type = str(_driver_data.m_session_type)
+            curr_lap_percent = (
+                f"{F1Utils.floatToStr(float(curr_lap_dist) / float(_driver_data.m_track_length) * 100.0)}%"
+                if curr_lap_dist is not None
+                else None
+            )
+
         else:
             lap_num = None
             sector = None
             curr_lap_time = None
             curr_lap_dist = None
 
-    with _globals_lock.gen_rlock():
-        if _globals.m_circuit is not None and _globals.m_event_type is not None:
-            track = _globals.m_circuit
-            event_type = _globals.m_event_type
-            curr_lap_percent = (
-                f"{F1Utils.floatToStr(float(curr_lap_dist) / float(_globals.m_packet_session.m_trackLength) * 100.0)}%"
-                if curr_lap_dist is not None
-                else None
-            )
-        else:
             track = None
             event_type = None
             curr_lap_percent = None
@@ -1936,8 +1213,8 @@ def isPositionHistorySupported() -> bool:
         bool: True if the position history is supported, False otherwise
     """
 
-    with _globals_lock.gen_rlock():
-        return bool(_globals.m_event_type and "Race" in _globals.m_event_type)
+    with _driver_data_lock.gen_rlock():
+        return _driver_data.isPositionHistorySupported()
 
 def processStreamUpdateButtonPress(custom_marker_obj: CustomMarkerEntry) -> None:
     """Processes the stream update button press event
@@ -1956,8 +1233,6 @@ def clearDataStructures(reason: str) -> None:
     """
     with _driver_data_lock.gen_wlock():
         _driver_data.clear()
-    with _globals_lock.gen_wlock():
-        _globals.clear()
     _custom_markers_history.clear()
     png_logger.debug(f"Clearing all data structures. Reason: {reason}")
 
@@ -1969,12 +1244,13 @@ def getTyreDeltaNotificationMessages() -> List[TyreDeltaMessage]:
     """
 
     # sourcery skip: assign-if-exp, extract-method
-    with _globals_lock.gen_rlock():
+    with _driver_data_lock.gen_rlock():
         # N/A for spectating or after race - maybe support this later
-        if _globals.m_is_spectating or _globals.m_packet_final_classification or _globals.m_event_type != "Time Trial":
+        if (_driver_data.m_globals.m_is_spectating or
+            _driver_data.m_globals.m_packet_final_classification or
+            _driver_data.m_globals.m_event_type != "Time Trial"):
             return []
 
-    with _driver_data_lock.gen_rlock():
         # If player ded, not applicable
         if (_driver_data.m_player_index is None) or (_driver_data.m_is_player_dnf):
             return []
